follow count bug

+++ app/models/organization.rb
@@ -6,7 +6,6 @@
 #  name                :string(255)
 #  description         :text
 #  email               :string(255)
-#  logo_url            :string(255)
 #  created_at          :datetime         not null
 #  updated_at          :datetime         not null
 #  avatar_file_name    :string(255)
@@ -17,7 +16,7 @@
 #
 
 class Organization < ActiveRecord::Base
-	has_attached_file :avatar, :styles => { :large => '120x120>', :medium => "100x100>", :thumb => "50x50>" }
+	has_attached_file :avatar, :styles => { :medium => "100x100>", :thumb => "50x50>" }
 	attr_accessible :description, :email, :name, :avatar, :category_id
 	
 	belongs_to :category
@@ -26,6 +25,9 @@ class Organization < ActiveRecord::Base
 	has_many :members, through: :member_relationships, source: :member
 	has_many :reverse_org_relationships, foreign_key: 'club_id', dependent: :destroy, class_name: 'OrgRelationship'
 	has_many :applicants, through: :reverse_org_relationships, source: :member
+	
+	has_many :club_admins, foreign_key: 'club_id', dependent: :destroy, class_name: 'ClubAdmin'
+	has_many :admins, through: :club_admins, source: :member
 	
 	validates :name, presence: true, length: { maximum: 30, minimum: 3 }, uniqueness: true
 	validates :email, presence: true, uniqueness: true
@@ -54,8 +56,24 @@ class Organization < ActiveRecord::Base
 	
 	def remove!(member)
 		mem_rel = has_member? member
-		if mem_rel
-			mem_rel.destroy
+		mem_rel.destroy unless !mem_rel
+	end
+	
+	def has_admin? user
+		return nil if !has_member? user
+		club_admins.find_by_member_id(user.id)
+	end
+	
+	def add_admin! user
+		if !has_member? user
+			user.join! self
+			accept! user
 		end
+		club_admins.create!(member_id: user.id) unless self.has_admin? user
+	end
+	
+	def remove_admin! user
+		admin = has_admin? user
+		admin.destroy unless !admin
 	end
 end
